requirements.txt
# Core dependencies
gymnasium>=0.29.0
gymnasium-minigrid>=1.0.0
numpy>=1.24.0
torch>=2.0.0
sentence-transformers>=2.2.0
matplotlib>=3.5.0
imageio>=2.25.0
imageio-ffmpeg>=0.4.8
Pillow>=9.0.0